debug

+++ app/controllers/admin/events_controller.rb
@@ -49,8 +49,19 @@ class Admin::EventsController < AdminController
     total = 0
     Array(params[:ids]).each do |event_id|
       event = Event.find(event_id)
-      event.destroy
-      total += 1
+      #event.destroy
+      #total += 1
+
+      if params[:commit] == I18n.t(:bulk_update)
+        event.status = params[:event_status]
+        if event.save
+          total += 1
+        end
+      elsif params[:commit] == I18n.t(:bulk_delete)
+        event.destroy
+        total += 1
+      end
+
     end
 
     flash[:alert] = "成功完成 #{total} 笔"
@@ -60,7 +71,6 @@ class Admin::EventsController < AdminController
   protected
 
   def event_params
-    #params.require(:event).permit(:name, :description, :friendly_id, :status, :category_id)
     params.require(:event).permit(:name, :description, :friendly_id, :status, :category_id, :tickets_attributes => [:id, :name, :description, :price, :_destroy])
   end
